rename raw-exec-plugin to raw-exec
