+++ plugins/drivers/raw-exec/main.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/hashicorp/go-plugin"
 
-	"github.com/hashicorp/nomad/plugins/drivers/raw-exec-plugin/proto"
-	"github.com/hashicorp/nomad/plugins/drivers/raw-exec-plugin/shared"
+	"github.com/hashicorp/nomad/plugins/drivers/raw-exec/proto"
+	"github.com/hashicorp/nomad/plugins/drivers/raw-exec/shared"
 )
 
 func main() {
